// MCP_Frontend-main/src/app/shared/models/loan-account.ts
export class LoanAccount {
    loanId: number;
    loanNumber: string;
    creditApplId: number;
    approvedAmount: number;
    startDate: Date;
    endDate: Date;
    ledgerBal: number;
    repayInterest: number;
    nextPmtDate: Date;
    nextPmtAmount: number;
    repayPrincipal: number;
    term: number;
    repayPeriod: string;
    status: string;
    createdBy: string;
    createDate: Date;
    creditNumber: string;
    custId: number;
    custName: string;
}