// MCP_Frontend-main/src/app/shared/models/loan-repayment.ts
export class LoanRepayment {
    transId: number;
    acctNo: string;
    txnRef: string;
    transAmount: number;
    chargeAmount: number;
    channelCode: string;
    postedBy: string;
    initiatorPhoneNo: string;
    transDescr: string;
    successFlag: string;
    transDate: Date;
    reversalFlag: string;
    reversalReason: string;
    externalTransRef: string;
}